+++ talaiot/src/main/kotlin/com/cdsap/talaiot/publisher/pushgateway/PushGatewayFormatter.kt
@@ -1,6 +1,7 @@
-package com.cdsap.talaiot.publisher
+package com.cdsap.talaiot.publisher.pushgateway
 
 import com.cdsap.talaiot.entities.ExecutionReport
+import com.cdsap.talaiot.publisher.formatTagPublisher
 
 /**
  * Formatter to format task and build execution data to Pushgateway metrics format.
