+++ talaiot/src/main/kotlin/com/cdsap/talaiot/publisher/pushgateway/PushGatewayPublisher.kt
@@ -1,8 +1,9 @@
-package com.cdsap.talaiot.publisher
+package com.cdsap.talaiot.publisher.pushgateway
 
 import com.cdsap.talaiot.configuration.PushGatewayPublisherConfiguration
 import com.cdsap.talaiot.entities.ExecutionReport
 import com.cdsap.talaiot.logger.LogTracker
+import com.cdsap.talaiot.publisher.Publisher
 import com.cdsap.talaiot.request.Request
 import java.util.concurrent.Executor
