+++ talaiot/src/test/kotlin/com/cdsap/talaiot/publisher/PushGatewayPublisherTest.kt
@@ -4,6 +4,8 @@ import com.cdsap.talaiot.configuration.PushGatewayPublisherConfiguration
 import com.cdsap.talaiot.entities.*
 
 import com.cdsap.talaiot.logger.TestLogTrackerRecorder
+import com.cdsap.talaiot.publisher.pushgateway.PushGatewayFormatter
+import com.cdsap.talaiot.publisher.pushgateway.PushGatewayPublisher
 import com.cdsap.talaiot.request.SimpleRequest
 import io.github.rybalkinsd.kohttp.dsl.httpGet
 import io.github.rybalkinsd.kohttp.ext.url
